requirements.txt
qdrant-client>=1.7.0
cohere>=4.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0